allow rolebased admin role change member status

+++ convex/files.ts
@@ -31,7 +31,8 @@ export const hasAccessToOrg = async (
   if (!user) return null;
 
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some((item) => item.orgId === orgId) ||
+    user.tokenIdentifier.includes(orgId);
 
   if (!hasAccess) {
     return null;
@@ -113,6 +114,14 @@ export const deleteFile = mutation({
     const access = await hasAccessToFile(ctx, args.fileId);
 
     if (!access) throw new ConvexError("no file access to file");
+
+    const isAdmin =
+      access.user.orgIds.find((org) => org.orgId === access.file.orgId)
+        ?.role === "admin";
+
+    if (!isAdmin) {
+      throw new ConvexError("Your have not admin access to delete");
+    }
 
     await ctx.db.delete(args.fileId);
   },
